Deleted problematic text filter
Deleted the regex expression that removed anything matching '\d\.' The original intent of this was to delete the numbering in numbered lists, but this had the unintended consequence of also truncating years, e.g. "the largest increase in productivity so far in 2016." would become "the largest increase in productivity so far in 201". In the end, this loss of information and coherence outweighs the original benefit of this particular filter.